made a new architecture and added docker files
