+++ internal/transort/handlers/managment.go
@@ -1,10 +1,10 @@
 package handlers
 
 import (
-	"net/http"
+	_ "github.com/ByteForge-Systems/vpn-api/internal/models"
+	"github.com/ByteForge-Systems/vpn-api/internal/node_client"
 	"github.com/gin-gonic/gin"
-	"github.com/ByteForge-Systems/vpn-api/client"
-	_"github.com/ByteForge-Systems/vpn-api/api/models"
+	"net/http"
 )
 
 // @Summary Перезапустить Xray
@@ -14,8 +14,9 @@ import (
 // @Success 200 {object} models.SuccessResponse
 // @Failure 500 {object} models.ErrorResponse
 // @Router /api/management/restart [post]
+
 func RestartXray(c *gin.Context) {
-	err := client.RestartXray()
+	err := node_client.RestartXray()
 	if err != nil {
 		c.JSON(http.StatusInternalServerError, gin.H{"error": err.Error()})
 		return
@@ -30,8 +31,9 @@ func RestartXray(c *gin.Context) {
 // @Produce json
 // @Success 200 {object} models.XrayStatus
 // @Router /api/management/status [get]
+
 func GetXrayStatus(c *gin.Context) {
-	status, err := client.GetXrayStatus()
+	status, err := node_client.GetXrayStatus()
 	if err != nil {
 		c.JSON(http.StatusInternalServerError, gin.H{"error": err.Error()})
 		return
@@ -46,13 +48,14 @@ func GetXrayStatus(c *gin.Context) {
 // @Success 200 {object} models.SuccessResponse
 // @Failure 500 {object} models.ErrorResponse
 // @Router /api/management/start [post]
+
 func StartXray(c *gin.Context) {
-    err := client.StartXray()
-    if err != nil {
-        c.JSON(http.StatusInternalServerError, gin.H{"error": err.Error()})
-        return
-    }
-    c.JSON(http.StatusOK, gin.H{"message": "Xray started"})
+	err := node_client.StartXray()
+	if err != nil {
+		c.JSON(http.StatusInternalServerError, gin.H{"error": err.Error()})
+		return
+	}
+	c.JSON(http.StatusOK, gin.H{"message": "Xray started"})
 }
 
 // @Summary Остановить Xray
@@ -62,11 +65,12 @@ func StartXray(c *gin.Context) {
 // @Success 200 {object} models.SuccessResponse
 // @Failure 500 {object} models.ErrorResponse
 // @Router /api/management/stop [post]
+
 func StopXray(c *gin.Context) {
-    err := client.StopXray()
-    if err != nil {
-        c.JSON(http.StatusInternalServerError, gin.H{"error": err.Error()})
-        return
-    }
-    c.JSON(http.StatusOK, gin.H{"message": "Xray stopped"})
+	err := node_client.StopXray()
+	if err != nil {
+		c.JSON(http.StatusInternalServerError, gin.H{"error": err.Error()})
+		return
+	}
+	c.JSON(http.StatusOK, gin.H{"message": "Xray stopped"})
 }
