+++ internal/transort/handlers/user.go
@@ -1,11 +1,11 @@
 package handlers
 
 import (
-	"net/http"
+	_ "github.com/ByteForge-Systems/vpn-api/internal/models"
+	"github.com/ByteForge-Systems/vpn-api/internal/node_client"
 	"github.com/gin-gonic/gin"
-	"github.com/ByteForge-Systems/vpn-api/client"
 	"github.com/google/uuid"
-	_ "github.com/ByteForge-Systems/vpn-api/api/models"
+	"net/http"
 )
 
 // @Summary Добавить пользователя
@@ -17,16 +17,17 @@ import (
 // @Success 200 {object} models.User
 // @Failure 400 {object} models.ErrorResponse
 // @Router /api/user/ [post]
-func AddUser(c *gin.Context) {
-    newUUID := uuid.New().String()
-    
-    _, err := client.AddUser(newUUID)
-    if err != nil {
-        c.JSON(http.StatusInternalServerError, gin.H{"error": err.Error()})
-        return
-    }
 
-    c.JSON(http.StatusOK, gin.H{"uuid": newUUID})
+func AddUser(c *gin.Context) {
+	newUUID := uuid.New().String()
+
+	_, err := node_client.AddUser(newUUID)
+	if err != nil {
+		c.JSON(http.StatusInternalServerError, gin.H{"error": err.Error()})
+		return
+	}
+
+	c.JSON(http.StatusOK, gin.H{"uuid": newUUID})
 }
 
 // @Summary Удалить пользователя
@@ -37,9 +38,10 @@ func AddUser(c *gin.Context) {
 // @Success 200 {object} models.SuccessResponse
 // @Failure 404 {object} models.ErrorResponse
 // @Router /api/user/{id} [delete]
+
 func RemoveUser(c *gin.Context) {
 	userID := c.Param("id")
-	err := client.RemoveUser(userID)
+	err := node_client.RemoveUser(userID)
 	if err != nil {
 		c.JSON(http.StatusInternalServerError, gin.H{"error": err.Error()})
 		return
@@ -54,15 +56,17 @@ func RemoveUser(c *gin.Context) {
 // @Produce json
 // @Success 200 {array} models.User
 // @Router /api/user/all [get]
-func GetAllUsers(c *gin.Context) {
-    clients, err := client.GetAllUsers() // слайс
-    if err != nil {
-        c.JSON(http.StatusInternalServerError, gin.H{"error": err.Error()})
-        return
-    }
 
-    c.JSON(http.StatusOK, gin.H{"users": clients})
+func GetAllUsers(c *gin.Context) {
+	clients, err := node_client.GetAllUsers() // слайс
+	if err != nil {
+		c.JSON(http.StatusInternalServerError, gin.H{"error": err.Error()})
+		return
+	}
+
+	c.JSON(http.StatusOK, gin.H{"users": clients})
 }
+
 // @Summary Сгенерировать VLESS-ссылку
 // @Description Генерирует VLESS-ссылку для пользователя по ID
 // @Tags User
@@ -71,9 +75,10 @@ func GetAllUsers(c *gin.Context) {
 // @Success 200 {object} models.VLESSLink
 // @Failure 404 {object} models.ErrorResponse
 // @Router /api/user/{id}/link [get]
+
 func GenerateVLESSLink(c *gin.Context) {
 	userID := c.Param("id")
-	link, err := client.GenerateVLESSLink(userID)
+	link, err := node_client.GenerateVLESSLink(userID)
 	if err != nil {
 		c.JSON(http.StatusInternalServerError, gin.H{"error": err.Error()})
 		return
